Replace static text with slots in Accordion.vue
Removed static texts from Accordion.vue component and replaced them with slots for a title and content. This allows the component to be more reusable across the application by adapting its content accordingly. Also updated the usage of the Accordion component in the AboutView.vue file to integrate these changes.