Update Sponsors component and add Partners component
Made a series of changes to the Sponsors.vue, including modifications to CSS classes and data. Also added a Partners.vue component which is almost like a clone of the Sponsors component. The new component is now imported and used in HomeView.vue. These changes bring more organization to the codebase and enhance the visual presentation for the users.