Increase logo size in Navbar and update image paths in Partners.vue
This commit increases the height of the logo in the Navbar.vue component, making it slightly more prominent. Additionally, the paths for partner images in the Partners.vue file have been updated. Instead of fetching from the 'assets' directory, they will now be retrieved from the 'img' directory in the public folder, aiming to improve resource organization and accessibility.